+++ fuel/packages/orm/classes/manymany.php
@@ -3,10 +3,10 @@
  * Fuel is a fast, lightweight, community driven PHP 5.4+ framework.
  *
  * @package    Fuel
- * @version    1.9-dev
+ * @version    1.8.2
  * @author     Fuel Development Team
  * @license    MIT License
- * @copyright  2010-2026 Fuel Development Team
+ * @copyright  2010 - 2019 Fuel Development Team
  * @link       https://fuelphp.com
  */
 
@@ -40,11 +40,6 @@ class ManyMany extends Relation
 
 	public function __construct($from, $name, array $config)
 	{
-		if (call_user_func_array(array($from, 'property'), array($name)))
-		{
-			throw new \FuelException(sprintf('"%s" is defined as both a property and a many-many relation in model %s', $name, $from));
-		}
-
 		$this->name        = $name;
 		$this->model_from  = $from;
 		$this->model_to    = array_key_exists('model_to', $config)
@@ -72,29 +67,10 @@ class ManyMany extends Relation
 		$this->key_through_to = ! empty($config['key_through_to'])
 			? (array) $config['key_through_to'] : (array) \Inflector::foreign_key($this->model_to);
 
-		// DEPRECATED SINCE 1.9
-		$this->cascade_delete  = array_key_exists('cascade_delete', $config)
-			? $config['cascade_delete'] : $this->cascade_delete;
-
-		if (array_key_exists('constraint', $config) and in_array($config['constraint'], $this->valid_constraints))
-		{
-			switch($config['constraint'])
-			{
-				case static::CONSTRAINT_RESTRICT:
-					$this->cascade_check = true;
-					break;
-				case static::CONSTRAINT_CASCADE:
-					$this->cascade_delete = true;
-					break;
-				case static::CONSTRAINT_SETDEFAULT:
-					$this->cascade_delete = false;
-					break;
-				default:
-			}
-		}
-
 		$this->cascade_save    = array_key_exists('cascade_save', $config)
 			? $config['cascade_save'] : $this->cascade_save;
+		$this->cascade_delete  = array_key_exists('cascade_delete', $config)
+			? $config['cascade_delete'] : $this->cascade_delete;
 
 		if ( ! class_exists($this->model_to))
 		{
@@ -366,39 +342,27 @@ class ManyMany extends Relation
 		}
 	}
 
-	public function delete($model_from, $parent_deleted, $cascade)
+	public function delete($model_from, $models_to, $parent_deleted, $cascade)
 	{
-		// fetch all related records
-		$model_from->get($this->name);
-
 		if ( ! $parent_deleted)
 		{
-			if ($this->cascade_check and ! empty($model_from->{$this->name}))
-			{
-				throw new \Orm\DeleteConstraintViolation($this->name);
-			}
-
 			return;
 		}
 
-		// break current relations, may be incomplete
+		// Remove relations
 		$model_from->unfreeze();
-
 		$rels = $model_from->_relate();
-		unset($rels[$this->name]);
+		$rels[$this->name] = array();
 		$model_from->_relate($rels);
-
 		$model_from->freeze();
 
 		// Delete all relationship entries for the model_from
 		$this->delete_related($model_from);
 
 		$cascade = is_null($cascade) ? $this->cascade_delete : (bool) $cascade;
-
-		if ($cascade)
+		if ($cascade and ! empty($models_to))
 		{
-			// yes, delete the reclated records
-			foreach ($model_from->{$this->name} as $m)
+			foreach ($models_to as $m)
 			{
 				$m->delete();
 			}
